% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{changedXyChr}
\alias{changedXyChr}
\title{Split chromosome X into pseudoautosomal region and non-pseudoautosomal region.}
\usage{
changedXyChr(plink, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{outputPrefix}{the prefix of the output PLINK format files.}
}
\value{
The output PLINK format files after splitting chromosome X into pseudoautosomal region and non-pseudoautosomal region.
}
\description{
Split chromosome X into pseudoautosomal region and non-pseudoautosomal region, if chromosome X data is available.
}
\author{
Junfang Chen <junfang.chen@zi-mannheim.de>
}
