% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasingImpute.R
\name{chunk4eachChr}
\alias{chunk4eachChr}
\title{Chunk each chromosome into multiple segments}
\usage{
chunk4eachChr(inputPrefix, outputPrefix, chrs, windowSize)
}
\arguments{
\item{inputPrefix}{the prefix of the input plink files for each chromosome.}

\item{outputPrefix}{the prefix of the output pure text files that keep all the chunks for each chromosome.}

\item{chrs}{specifiy the chromosomes for chunking.}

\item{windowSize}{the window size of each chunk.}
}
\value{
The output pure text files that keep all the chunks for each chromosome.
}
\description{
Chunk each chromosome into multiple segments by a predefined window size.
}
\author{
Junfang Chen <junfang.chen@zi-mannheim.de>
}
