% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasingImpute4.R
\name{computeInfoByQctool}
\alias{computeInfoByQctool}
\title{Calculate the info score by QCTOOL}
\usage{
computeInfoByQctool(qctool, inputSuffix, outputInfoFile)
}
\arguments{
\item{qctool}{an executable program in either the current working 
directory or somewhere in the command path.}

\item{inputSuffix}{the suffix of input .GEN files within current directory.}

\item{outputInfoFile}{the output info scores file consisting of 
two columns: all SNPs from .GEN files and their info scores.}
}
\value{
A pure text file contains the info scores of all SNPs from .GEN  
files with two columns: SNP names and the corresponding info scores.
}
\description{
Calculate the info score by QCTOOL for a set of SNPs from .GEN files.
}
\details{
These .GEN files may come from the output of impute4 results.
The intermediate generated files are retained in the directory.
}
\examples{
## In the current working directory
inputSuffix <- "noINDEL.gen"
outputInfoFile <- "infoScore.txt"
## computeInfoByQctool(qctool, inputSuffix, outputInfoFile)
}
\author{
Junfang Chen
}
