% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasingImpute.R
\name{.filterImputeData2}
\alias{.filterImputeData2}
\title{Filter genetic variants}
\usage{
.filterImputeData2(
  plink,
  outputInfoFile,
  infoScore = 0.6,
  inputPrefix,
  outputPrefix
)
}
\arguments{
\item{plink}{an executable program in either the current working 
directory or somewhere in the command path.}

\item{outputInfoFile}{the output file of info scores consisting of 
two columns: SNP names and their info scores. The headers are 
c("rsid", "info").}

\item{infoScore}{the cutoff of filtering imputation quality score for 
each variant. The default value is 0.6.}

\item{inputPrefix}{the prefix of the input imputed PLINK binary files.}

\item{outputPrefix}{the prefix of the output filtered PLINK binary files.}
}
\value{
A pure text file contains the info scores of all imputed SNPs with 
two columns: SNP names and the corresponding info scores. 
A pure text file with all excluded SNPs having bad info scores. 
The filtered PLINK binary imputed files,
}
\description{
Filter out genetic variants accoring to the info score.
}
\details{
Filter genetic variants accoring to the imputation quality score 
with the help of .impute2_info files generated by IMPUTE2. 
Often, we keep variants with imputation info score of greater than 0.6.    
Note that imputed SNPs with more than two alleles are not considered.
}
\author{
Junfang Chen
}
