% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasingImpute.R
\name{.getInfoScoreImpute2}
\alias{.getInfoScoreImpute2}
\title{Extract info score}
\usage{
.getInfoScoreImpute2(suffix4impute2info, outputInfoFile)
}
\arguments{
\item{suffix4impute2info}{the suffix of input IMPUTE2 generated files that 
store the imputation quality score for each variant from .impute2_info files.}

\item{outputInfoFile}{the output file of impute2 info scores consisting of 
two columns: all imputed SNPs and their info scores.}
}
\value{
A pure text file contains the info scores of all imputed SNPs with 
two columns: SNP names and the corresponding info scores. The header names 
are "rsid" and "info".
}
\description{
Extract info score file summarized from the output of IMPUTE2.
}
\author{
Junfang Chen
}
