% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasingImpute.R
\name{.imputedByImpute2}
\alias{.imputedByImpute2}
\title{Impute genotypes using IMPUTE2}
\usage{
.imputedByImpute2(impute2, chrs, prefixChunk, phaseDIR, referencePanel,
  impRefDIR, imputedDIR, prefix4eachChr, nCore, effectiveSize = 20000)
}
\arguments{
\item{impute2}{an executable program in either the current 
working directory or somewhere in the command path.}

\item{chrs}{specifiy the chromosome codes for imputation.}

\item{prefixChunk}{the prefix of the chunk files for each chromosome, 
along with the proper location directory.}

\item{phaseDIR}{the directory where prephased haplotypes are located.}

\item{referencePanel}{a string indicating the type of imputation 
reference panels is used: c("1000Gphase1v3_macGT1", "1000Gphase3").}

\item{impRefDIR}{the directory where the imputation reference files 
are located.}

\item{imputedDIR}{the directory where imputed files will be located.}

\item{prefix4eachChr}{the prefix of IMPUTE2 files for each chunk.}

\item{nCore}{the number of cores used for computation.}

\item{effectiveSize}{this parameter controls the effective population size.
Commonly denoted as Ne. A universal -Ne value of 20000 is suggested.}
}
\value{
The imputed files for all chunks from the given chromosomes.
}
\description{
Perform imputation by IMPUTE2 for the autosomal and sex chromosome 
prephased known haplotypes with a reference panel.
}
\details{
If ChrX is available then it is done automatically by passing 
the flag --chrX to IMPUTE2.
}
\seealso{
\code{\link{phaseImpute2}}.
}
\author{
Junfang Chen
}
