% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasingImpute4.R
\name{.imputedByImpute4}
\alias{.imputedByImpute4}
\title{Impute genotypes using IMPUTE4}
\usage{
.imputedByImpute4(impute4, chrs, prefixChunk, phaseDIR, referencePanel,
  impRefDIR, imputedDIR, prefix4eachChr, nCore, effectiveSize = 20000)
}
\arguments{
\item{impute4}{an executable program in either the current 
working directory or somewhere in the command path.}

\item{chrs}{specifiy the chromosome codes for imputation.}

\item{prefixChunk}{the prefix of the chunk files for each chromosome, 
along with the proper location directory.}

\item{phaseDIR}{the directory where prephased haplotypes are located.}

\item{referencePanel}{a string indicating the type of imputation 
reference panels is used: c("1000Gphase1v3_macGT1", "1000Gphase3").}

\item{impRefDIR}{the directory where the imputation reference files 
are located.}

\item{imputedDIR}{the directory where imputed files will be located.}

\item{prefix4eachChr}{the prefix of IMPUTE2 files for each chunk.}

\item{nCore}{the number of cores used for computation.}

\item{effectiveSize}{this parameter controls the effective population size.
Commonly denoted as Ne. A universal -Ne value of 20000 is suggested.}

\item{XPAR}{a logical value indicating whether --chrX flag should be 
passed for prephasing using SHAPEIT.
--chrX flag, specifically for chrX imputation'}
}
\value{
The imputed files for all chunks from given chromosomes, except   
sex chromosomes.
}
\description{
Perform imputation by IMPUTE4 for the autosomal prephased known haplotypes 
with a reference panel.
}
\seealso{
\code{\link{phaseImpute4}}.
}
\author{
Junfang Chen
}
