% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align2reference.R
\name{.prepareLegend2bim}
\alias{.prepareLegend2bim}
\title{Prepare a bim-like reference file}
\usage{
.prepareLegend2bim(inputFile, referencePanel, outputFile, ncore = 20)
}
\arguments{
\item{inputFile}{a set of legend file from the imputation reference panel.
For example, 1000 genome projects.}

\item{referencePanel}{a string indicating the type of imputation 
reference panels is used: c("1000Gphase1v3_macGT1", "1000Gphase3").}

\item{outputFile}{the pure text file that stores the prepared PLINK BIM
file alike format data.}

\item{ncore}{the number of cores used for computation.
The default is 20.}
}
\value{
Prepare a bim-like reference file. Note that the column names
are already defined, i.e. "chr", "rsID", "pos", "a0", "a1."
}
\description{
Prepare a bim-like file from the imputation reference legend file
(e.g. 1000 genome project).
}
\details{
To prepare a bim-like reference file from legend files.
One should first extract the specific content from these legend files
after downloading. Note that extract only biallelic SNPs (only 1 allele
in column3 and 4, and start with 'rs") and remove duplicated snp IDs.
Column names are added in the end.
}
\author{
Junfang Chen
}
