% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasingImpute.R
\name{.prePhasingByShapeit}
\alias{.prePhasingByShapeit}
\title{Prephasing genotypes using SHAPEIT}
\usage{
.prePhasingByShapeit(
  shapeit,
  chrs,
  dataDIR,
  prefix4eachChr,
  referencePanel,
  impRefDIR,
  phaseDIR,
  nThread = 40,
  effectiveSize = 20000,
  nCore = 1
)
}
\arguments{
\item{shapeit}{an executable program in either the 
current working directory or somewhere in the command path.}

\item{chrs}{specifiy the chromosome codes for phasing.}

\item{dataDIR}{the directory where genotype PLINK binary files are located.}

\item{prefix4eachChr}{the prefix of PLINK binary files for 
each chromosome.}

\item{referencePanel}{a string indicating the type of imputation 
reference panels is used: c("1000Gphase1v3_macGT1", "1000Gphase3").}

\item{impRefDIR}{the directory where the imputation reference files 
are located.}

\item{phaseDIR}{the directory where resulting pre-phased files 
will be located.}

\item{nThread}{the number of threads used for computation.
The default value is 40.}

\item{effectiveSize}{this parameter controls the effective population size. 
The default value is 20000.}

\item{nCore}{the number of cores used for computation. This can be tuned 
along with nThread. The default value is 1.}
}
\value{
The pre-phased haplotypes for given chromosomes.
}
\description{
Perform prephasing for study genotypes by SHAPEIT for the autosomal and 
sex chromosome haplotypes using a reference panel (pre-set).
}
\details{
If ChrX is available then it is done automatically by passing 
the flag --chrX to SHAPEIT.
}
\seealso{
\code{\link{phaseImpute2}}.
}
\author{
Junfang Chen
}
