% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasingImpute.R
\name{filterImputeData}
\alias{filterImputeData}
\title{Filter genetic variants}
\usage{
filterImputeData(plink, suffix4impute2info, outputInfoFile, infoScore = 0.6,
  badImputeSNPfile, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working 
directory or somewhere in the command path.}

\item{suffix4impute2info}{the suffix of input IMPUTE2 generated files that 
store the imputation quality score for each variant from .impute2_info files.}

\item{outputInfoFile}{the output file of impute2 info scores consisting of 
two columns: all imputed SNPs and their info scores.}

\item{infoScore}{the cutoff of filtering imputation quality score for 
each variant. The default value is 0.6.}

\item{badImputeSNPfile}{the output file of SNPs with bad info scores.}

\item{inputPrefix}{the prefix of the input imputed PLINK binary files.}

\item{outputPrefix}{the prefix of the output filtered PLINK binary files.}
}
\value{
A pure text file contains the info scores of all imputed SNPs with 
two columns: SNP names and the corresponding info scores. 
A pure text file with all excluded SNPs having bad info scores. 
The filtered PLINK binary imputed files,
}
\description{
Filter out genetic variants accoring to the imputation quality score.
}
\details{
Filter genetic variants accoring to the imputation quality score with 
the help of .impute2_info files generated by IMPUTE2. 
Often, we keep variants with imputation info score of greater than 0.6.    
Note that imputed SNPs with more than two alleles are not considered.
}
\author{
Junfang Chen
}
