% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasingImpute.R
\name{formatConvertGtool}
\alias{formatConvertGtool}
\title{Convert IMPUTE2 format files into PLINK format}
\usage{
formatConvertGtool(gtool, chrs, prefixChunk, phaseDIR, imputedDIR,
  prefix4plinkEachChr, suffix4imputed, postImputeDIR, nCore)
}
\arguments{
\item{gtool}{an executable GTOOL program in either the current working directory or somewhere in the command path.}

\item{chrs}{specifiy the chromosomes for conversion}

\item{prefixChunk}{the prefix of the chunk files for each chromosome, along with the proper location directory.}

\item{phaseDIR}{the directory where pre-phased files are located.}

\item{imputedDIR}{the directory where the imputated files are located.}

\item{prefix4plinkEachChr}{the prefix of the input IMPUTE2 files and also the output PLINK files for each chunk.}

\item{suffix4imputed}{the suffix of the IMPUTE2 format file that stores the imputed value.}

\item{postImputeDIR}{the directory where converted PLINK files will be located.}

\item{nCore}{the number of cores used for computation.}
}
\value{
The converted PLINK format files for each chunk.
}
\description{
Convert all chunks of IMPUTE2 format files into PLINK format using GTOOL.
}
\author{
Junfang Chen <junfang.chen@zi-mannheim.de>
}
