% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managePlinkData.R
\name{getGroupLabel}
\alias{getGroupLabel}
\title{Get the outcome label of the genotype data}
\usage{
getGroupLabel(inputFAMfile)
}
\arguments{
\item{inputFAMfile}{the PLINK FAM file.}
}
\value{
The group label of the genotype data: "control" or "case" or 
"caseControl" indicating both groups exist.
}
\description{
Get the group label from the PLINK FAM file.
}
\details{
If the input FAM file also has missing outcomes, which are 
shown in the sixth column of FAM file as "0", then an error is given.
}
\examples{
famFile <- system.file("extdata", "genoUpdatedData.fam", package="Gimpute")  
getGroupLabel(inputFAMfile=famFile)
}
\author{
Junfang Chen
}
