% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputedByImpute2.R
\name{imputedByImpute2}
\alias{imputedByImpute2}
\title{Impute genotypes using IMPUTE2}
\usage{
imputedByImpute2(impute2, chrs, prefixChunk, phaseDIR, impRefDIR, imputedDIR,
  prefix4plinkEachChr, nCore, effectiveSize)
}
\arguments{
\item{impute2}{an executable IMPUTE2 program in either the current working directory or somewhere in the command path.}

\item{chrs}{specifiy the chromosomes for imputation.}

\item{prefixChunk}{the prefix of the chunk files for each chromosome, along with the proper location directory.}

\item{phaseDIR}{the directory where prephased haplotypes are located.}

\item{impRefDIR}{the directory where the imputation reference files are located.}

\item{imputedDIR}{the directory where imputed files will be located.}

\item{prefix4plinkEachChr}{the prefix of IMPUTE2 files for each chunk.}

\item{nCore}{the number of cores used for computation.}

\item{effectiveSize}{this parameter controls the effective population size.}

\item{XPAR}{a logical value indicating whether --chrX flag should be passed for prephasing using SHAPEIT.
--chrX flag, specifically for chrX imputation'}
}
\value{
The imputed files for all chunks from given chromosomes.
}
\description{
Perform imputation by IMPUTE2 for the autosomal and sex chromosome prephased known haplotypes with a reference panel.
}
\author{
Junfang Chen <junfang.chen3@gmail.com>
}
