% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeImputeData.R
\name{mergeImputeData}
\alias{mergeImputeData}
\title{Merge chunk-wise PLINK files}
\usage{
mergeImputeData(plink, chrs, prefix4plinkEachChr, prefix4imputedPlink, nCore)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{chrs}{specifiy the chromosomes to be merged.}

\item{prefix4plinkEachChr}{the prefix of the input chunk-wise IMPUTE2 files.}

\item{prefix4imputedPlink}{the prefix of the final imputed PLINK format files.}

\item{nCore}{the number of cores used for computation.}
}
\value{
The merged PLINK format files.
}
\description{
Merge all chunk-wise PLINK files into chromosome-wise PLINK files then assemble into one PLINK file set. 
Create a file containing a list chunk ped and map file names.
At last, combine all chrs (combine the first 23 chrs; then Xpar).
}
\author{
Junfang Chen <junfang.chen3@gmail.com>
}
