% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasingImpute.R
\name{mergePlinkData}
\alias{mergePlinkData}
\title{Merge chunk-wise PLINK files}
\usage{
mergePlinkData(plink, chrs, prefix4plinkEachChr, prefix4mergedPlink, nCore)
}
\arguments{
\item{plink}{an executable program in either the current working 
directory or somewhere in the command path.}

\item{chrs}{specifiy the chromosome codes to be merged.}

\item{prefix4plinkEachChr}{the prefix of the input chunk-wise PLINK 
files.}

\item{prefix4mergedPlink}{the prefix of the final output PLINK 
binary files.}

\item{nCore}{the number of cores used for computation.}
}
\value{
The merged genome-wide PLINK binary files.
}
\description{
Merge all chunk-wise PLINK binary files into chromosome-wise PLINK 
binary files then assemble into a genome-wide PLINK binary file set.
}
\details{
Create a file containing a list chunk-wise PLINK PED and MAP 
file names. The prefix of these files must already indicate in which 
chromosome they belong to and files from the same chromosome will be 
combined. Then all chromosomal PLINK files are assembled together 
into one whole genome PLINK binary file set.
}
\author{
Junfang Chen
}
