% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasingImpute.R
\name{phaseImpute2}
\alias{phaseImpute2}
\title{Phasing and imputation}
\usage{
phaseImpute2(inputPrefix, outputPrefix, prefix4final, plink, shapeit, impute2,
  gtool, windowSize = 3e+06, effectiveSize = 20000, nCore4phase = 1,
  nThread = 40, nCore4impute = 40, nCore4gtool = 40, infoScore = 0.6,
  outputInfoFile, impRefDIR, tmpImputeDir, keepTmpDir = TRUE)
}
\arguments{
\item{inputPrefix}{the prefix of the input PLINK binary files for 
the imputation.}

\item{outputPrefix}{the prefix of the output PLINK binary files
after imputation and filtering out bad imputed variants.}

\item{prefix4final}{the prefix of the output PLINK binary files
after imputation.}

\item{plink}{an executable program in either the current 
working directory or somewhere in the command path.}

\item{shapeit}{an executable program in either the current 
working directory or somewhere in the command path.}

\item{impute2}{an executable program in either the current 
working directory or somewhere in the command path.}

\item{gtool}{an executable program in either the current 
working directory or somewhere in the command path.}

\item{windowSize}{the window size of each chunk. 
The default value is 3000000.}

\item{effectiveSize}{this parameter controls the effective population size. 
The default value is 20000.}

\item{nCore4phase}{the number of cores used for phasing. This can be tuned 
along with nThread. The default value is 1}

\item{nThread}{the number of threads used for computation.
The default value is 40.}

\item{nCore4impute}{the number of cores used for imputation. 
The default value is 40.}

\item{nCore4gtool}{the number of cores used for computation. 
The default value is 40.}

\item{infoScore}{the cutoff of filtering imputation quality score for 
each variant. The default value is 0.6.}

\item{outputInfoFile}{the output file of impute2 info scores consisting of 
two columns: all imputed SNPs and their info scores.}

\item{impRefDIR}{the directory where the imputation reference files 
are located.}

\item{tmpImputeDir}{the name of the temporary directory used for 
storing phasing and imputation results.}

\item{keepTmpDir}{a logical value indicating if the directory 
'tmpImputeDir' should be kept or not. The default is TRUE.}
}
\value{
1.) The filtered imputed PLINK binary files; 
2.) The final PLINK binary files including bad imputed variants;
3.) A pure text file contains the info scores of all imputed SNPs with 
two columns: SNP names and the corresponding info scores.
}
\description{
Perform phasing, imputation and conversion from IMPUTE2 format into PLINK 
binary files.
}
\details{
The whole imputation process mainly consists of the following 
steps: 
1.) Phasing the input PLINK data using an existing imputation reference;  
2.) Imputing the input PLINK data using phased results and an existing 
reference data; 
3.) Converting IMPUTE2 format data into PLINK format.
4.) Combining all imputed data into whole-genome PLINK binary files.
5.) Filtering out imputed variants with bad imputation quality.
}
\examples{
## In the current working directory
bedFile <- system.file("extdata", "removedMonoSnp.bed", package="Gimpute")
bimFile <- system.file("extdata", "removedMonoSnp.bim", package="Gimpute") 
famFile <- system.file("extdata", "removedMonoSnp.fam", package="Gimpute")
system(paste0("scp ", bedFile, bimFile, famFile, " ."))   
inputPrefix <- "removedMonoSnp"  
outputPrefix <- "gwasImputedFiltered"
prefix4final <- "gwasImputed"   
outputInfoFile <- "impute2infoUpdated.txt"
tmpImputeDir <- "tmpImpute"
## Not run: Requires an executable program PLINK, e.g.
## plink <- "/home/tools/plink"
phaseImpute2(inputPrefix, outputPrefix, prefix4final,
            plink, shapeit, impute2, gtool, 
            windowSize=3000000, effectiveSize=20000, 
            nCore4phase=1, nThread=40, 
            nCore4impute=40, nCore4gtool=40, 
            infoScore=0.6, outputInfoFile, 
            impRefDIR, tmpImputeDir, keepTmpDir=TRUE)
}
\references{
\enumerate{
  \item Howie, B., et al. (2012). Fast and accurate genotype imputation 
        in genome-wide association studies through pre-phasing. Nat Genet 
        44(8): 955-959.
  \item Howie, B. N., et al. (2009). A flexible and accurate genotype 
        imputation method for the next generation of genome-wide association 
        studies. PLoS Genet 5(6): e1000529.
}
}
\author{
Junfang Chen
}
