% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postImputation.R
\name{postImpQC}
\alias{postImpQC}
\title{Post imputation quality control}
\usage{
postImpQC(inputPrefix, out1, out2, out3, out4, out5, out6, outputInfoFile,
  infoScore = 0.6, inputPrefix4aligned2impRef, missCutoff = 20,
  outRemovedSNPfile, outRetainSNPfile, referencePanel)
}
\arguments{
\item{inputPrefix}{the prefix of the final imputed PLINK files.}

\item{out2}{the prefix of the final imputed PLINK files with the index.}

\item{out3}{the prefix of well imputed PLINK files with the index.}

\item{out4}{the prefix of well imputed PLINK files after removing any 
SNPs with the same positions (if any), the index is also appended.}

\item{out5}{the prefix of well imputed PLINK files with the index 
after adding previously identified monomorphic SNPs if any.}

\item{out6}{the prefix of final well imputed PLINK files with the index.}

\item{outputInfoFile}{the output file of impute2 info scores consisting of 
two columns: all imputed SNPs and their info scores.}

\item{infoScore}{the cutoff of filtering imputation quality score for 
each variant. The default value is 0.6.}

\item{inputPrefix4aligned2impRef}{the prefix of the output PLINK binary 
files after removing SNPs whose alleles are not in the imputation reference,
taking their genomic positions into account.}

\item{missCutoff}{the cutoff of the least number of instances for 
a SNP that is not missing. The default is 20.}

\item{outRemovedSNPfile}{the output file of SNPs with pre-defined 
missing values that are removed.}

\item{outRetainSNPfile}{the output file of SNPs that are retained.}

\item{referencePanel}{a string indicating the type of imputation 
reference panels is used: c("1000Gphase1v3_macGT1", "1000Gphase3").}
}
\value{
All imputed genotype data in PLINK format, well imputed data and 
its variants, as well as a set of pure text files containing removed or 
retained SNPs.
}
\description{
Perform quality control and data management after imputation.
}
\author{
Junfang Chen
}
