% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{prepareAnnoFile4affy}
\alias{prepareAnnoFile4affy}
\title{prepare Affymetrix chip annotation file}
\usage{
prepareAnnoFile4affy(inputFile, outputFile, chipType)
}
\arguments{
\item{inputFile}{an input pure text file that contains the chip annotation
information.}

\item{outputFile}{an output pure text file that stores the chip annotation 
information in a user-defined format.}

\item{chipType}{a string name defines the type of the chip annotation file: 
'SNPIDstudy', and 'rsIDstudy'.}
}
\value{
a pure text file that stores the prepared chip annotation 
information in a user-defined format.
}
\description{
Prepare Affymetrix chip annotation file into the format of interest.
}
\details{
If the chip annotation file is not available for your study, 
it can be downloaded from http://www.well.ox.ac.uk/~wrayner/strand/.
The chip annotation file is organized into two different types: 
\enumerate{
  \item If the snp name of your study genotype data starts with "SNP_", 
        then the chip type "SNPIDstudy" is used; Usually, Affymetrix chip 
        data belongs to this category. The prepared output annotation file  
        must at least consist of the following column names: 
        SNPIDstudy, rs, chr, pos, strand.
  \item If the snp name of your study genotype data starts with "rs", then
        the chip type "rsIDstudy" is used; The prepared output annotation 
        file must at least consist of the following column names:  
        SNPIDstudy, rs, chr, pos, strand. Illumina chip is often specified 
        in this format.
}
The column "strand" must only have two kinds of values "-" and "+". 
Variants with all other values should be excluded.
}
\author{
Junfang Chen
}
