% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{prepareChipAnnoFile4PsychChip}
\alias{prepareChipAnnoFile4PsychChip}
\title{prepare PsychChip chip annotation file}
\usage{
prepareChipAnnoFile4PsychChip(inputFile, outputFile)
}
\arguments{
\item{inputFile}{an input pure text file that contains the chip annotation
information.}

\item{outputFile}{an output pure text file that stores the chip annotation 
information in a pre-defined format.}
}
\value{
a pure text file that stores the prepared chip annotation 
information in a user-defined format.
}
\description{
Prepare PsychChip chip annotation file into the format of interest.
}
\details{
If the chip annotation file is not available for your study, it can be 
downloaded from http://www.well.ox.ac.uk/~wrayner/strand/.
}
\author{
Junfang Chen
}
