% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postImputation.R
\name{reductExpand}
\alias{reductExpand}
\title{Post imputation data extraction and expansion}
\usage{
reductExpand(referencePanel, inputPrefix, inputQCprefix, snpRefAlleleFile,
  snpDiffAlleleFile, snpMissPosFile, snpSameNameDifPosFile, reducedToSpecificfn,
  specificDiffAllelefn, specificMissPosfn, specificDiffPosfn)
}
\arguments{
\item{referencePanel}{a string indicating the type of imputation 
reference panels is used: c("1000Gphase1v3_macGT1", "1000Gphase3").}

\item{inputPrefix}{the prefix of final well imputed PLINK files.}

\item{inputQCprefix}{the prefix of QCed PLINK files.}

\item{snpRefAlleleFile}{the output plain text file that stores the
removed SNPs whose alleles are in the imputation reference,
taking their genomic positions into account.}

\item{snpDiffAlleleFile}{the output plain text file that stores the removed 
SNPs whose alleles are not in the imputation reference,
taking their genomic positions into account.}

\item{snpMissPosFile}{the output plain text file that stores the removed SNPs
whose genomic positions are not in the imputation reference,
ingoring SNP names.}

\item{snpSameNameDifPosFile}{the output plain text file that stores the 
removed SNPs whose genomic positions are not in the imputation reference,
taking SNP names into account.}

\item{reducedToSpecificfn}{the prefix of PLINK files which are reduced in the 
number of SNPs and contain only SNPs prior to imputation.}

\item{specificDiffAllelefn}{the prefix of PLINK files which added genotypes
with different alleles other than the imputation to the file with the 
prefix: reducedToSpecificfn.}

\item{specificMissPosfn}{the prefix of PLINK files which added genotypes
with missing positions other than the imputation to the file with the 
prefix: specificDiffAllelefn.}

\item{specificDiffPosfn}{the prefix of PLINK files which added genotypes
with different positions other than the imputation to the file with the 
prefix: specificMissPosfn .}
}
\value{
Extracted genotypes from the imputed data which contain SNPs before
imputation. On the basis of extracted genotypes, genotypes different from 
the imputation reference panel are appended, including SNPs with different  
alleles, missing positions, and different positions.
}
\description{
Reduce well imputed dataset to have SNPs before imputation and then added 
genotype data that are different from the imputation reference panel.
}
\author{
Junfang Chen
}
