% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{removedDoubleProbes}
\alias{removedDoubleProbes}
\title{Remove duplicated probes or SNPs}
\usage{
removedDoubleProbes(plink, inputPrefix, chipAnnoFile, chipType,
  outputSNPdupFile, outputPrefix)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{chipAnnoFile}{a pure text file that stores the chip annotation information (Affymetrix, Illumination, PsychChip and so on). 
This file can be found http://www.well.ox.ac.uk/~wrayner/strand/.}

\item{chipType}{a string name that defines the type of the chip annotation file: 'Illumina', 'affymetrix' or 'psychChip'.}

\item{outputSNPdupFile}{a pure text file that stores the duplicated probe/SNP IDs, which can be found with the help of the chip annotation file.}

\item{outputPrefix}{the prefix of the output PLINK format files.}
}
\value{
The output PLINK format files after removing duplicated probe IDs.
}
\description{
Remove duplicated probes or SNPs that have same rs-names but different versions of SNP ID (e.g. SNP-A IDs for Affymetrix chip) found in chip annotation information (For Affymetrix chip and PsychChip). 
This chip annotation file is defined in the configuration folder.
}
\author{
Junfang Chen <junfang.chen@zi-mannheim.de>
}
