% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{removedExclProbe}
\alias{removedExclProbe}
\title{Remove improper SNPs}
\usage{
removedExclProbe(plink, inputPrefix, excludedProbeIdsFile, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{excludedProbeIdsFile}{a pure text file that stores the SNP IDs, 
one per line, which need to be removed. If it is null, then skip this step.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files after removing unwanted SNP IDs.
}
\description{
Remove SNPs that may be duplicated, or with unexpected SNP names.
}
\details{
excludedProbeIdsFile should be defined in a plain text file 
in advance. Improper SNPs such as AFFX, cnvi etc.. with unexpect format
must be excluded.
}
\author{
Junfang Chen
}
