% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeQC.R
\name{removedInstFhet}
\alias{removedInstFhet}
\title{Remove subjects with abnormal autosomal heterozygosity deviation}
\usage{
removedInstFhet(plink, Fhet, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{Fhet}{the cutoff of the autosomal heterozygosity deviation.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files after removing subjects with great 
autosomal heterozygosity deviation.
}
\description{
Remove subjects with great autosomal heterozygosity deviation.
}
\details{
If the cutoff of the autosomal heterozygosity deviation is set to 
be greater than 0.2, i.e. |Fhet| >= 0.2, then this analysis will 
automatically skip haploid markers (male X and Y chromosome markers).
}
\examples{
## In the current working directory
bedFile <- system.file("extdata", "genoUpdatedData.bed", package="Gimpute")
bimFile <- system.file("extdata", "genoUpdatedData.bim", package="Gimpute") 
famFile <- system.file("extdata", "genoUpdatedData.fam", package="Gimpute")
system(paste0("scp ", bedFile, bimFile, famFile, " ."))  
Fhet <- 0.2
inputPrefix <- "genoUpdatedData" 
outputPrefix <- "2_06_removedInstFhet" 
## Not run: Requires an executable program PLINK, e.g.
## plink <- "/home/tools/plink"
## removedInstFhet(plink, Fhet, inputPrefix, outputPrefix)
}
\author{
Junfang Chen
}
