% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removedInstMiss.R
\name{removedInstMiss}
\alias{removedInstMiss}
\title{Remove subjects with missing values}
\usage{
removedInstMiss(plink, sampleMissCutOff, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{sampleMissCutOff}{the cutoff of the missingness for removing subjects/instances.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{outputPrefix}{the prefix of the output PLINK format files.}
}
\value{
The output PLINK format files after removing subjects with pre-defined missing values.
}
\description{
Remove Subjects or instances with missingness of greater than a certain threshold.
}
\author{
Junfang Chen <junfang.chen3@gmail.com>
}
