% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeQC.R
\name{removedParentIdsMiss}
\alias{removedParentIdsMiss}
\title{Reset paternal and maternal codes}
\usage{
removedParentIdsMiss(plink, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files.
}
\description{
Reset paternal and maternal codes of non-founders if parents not present. 
Replace the paternal ID and maternal ID of subjects (childs) by the
value zero if the paternal ID and the maternal ID do not belong to any
subject (parent) with the same family ID as the child.
}
\details{
By default, if parental IDs are provided for a sample, 
they are not treated as a founder even if neither parent is 
in the dataset.  With no modifiers, --make-founders clears 
both parental IDs whenever at least one parent is not in the dataset, 
and the affected samples are now considered founders.
}
\examples{
## In the current working directory
bedFile <- system.file("extdata", "genoUpdatedData.bed", package="Gimpute")
bimFile <- system.file("extdata", "genoUpdatedData.bim", package="Gimpute") 
famFile <- system.file("extdata", "genoUpdatedData.fam", package="Gimpute")
system(paste0("scp ", bedFile, bimFile, famFile, " ."))  
inputPrefix <- "genoUpdatedData" 
outputPrefix <- "2_07_removedParentIdsMiss" 
## Not run: Requires an executable program PLINK, e.g.
## plink <- "/home/tools/plink"
## removedParentIdsMiss(plink, inputPrefix, outputPrefix)
}
\author{
Junfang Chen
}
