% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removedSnpFemaleChrXhweAll.R
\name{removedSnpFemaleChrXhweAll}
\alias{removedSnpFemaleChrXhweAll}
\title{Hardy weinberg equilibrium test for chromosome X SNPs in females.}
\usage{
removedSnpFemaleChrXhweAll(plink, inputPrefix, pval, outputFile_pVal,
  outputSNPfile, outputPrefix)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{pval}{the p-value cutoff for controlling HWE test in female subjects. Only chromosome X SNPs are tested here.}

\item{outputFile_pVal}{the output pure text file that stores chromosome X SNPs and their sorted HWE p-values.}

\item{outputSNPfile}{the output pure text file that stores the removed SNPs.}

\item{outputPrefix}{the prefix of the output PLINK format files.}
}
\value{
The output PLINK format files.
}
\description{
Hardy weinberg equilibrium test for chromosome X SNPs in all female subjects. 
 If only control data, then it's the same as the test with all female control subjects.
}
\author{
Junfang Chen <junfang.chen3@gmail.com>
}
