% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeQC.R
\name{removedSnpFemaleChrXhweControl}
\alias{removedSnpFemaleChrXhweControl}
\title{Hardy Weinberg Equilibrium test for chromosome X SNPs in female controls.}
\usage{
removedSnpFemaleChrXhweControl(plink, inputPrefix, pval, outputPvalFile,
  outputSNPfile, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{pval}{the p-value cutoff for controlling HWE test in female control 
subjects. Only chromosome X SNPs are considered.}

\item{outputPvalFile}{the output pure text file that stores chromosome X SNPs 
and their sorted HWE p-values.}

\item{outputSNPfile}{the output pure text file that stores the removed SNPs, 
one per line.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files.
}
\description{
Hardy Weinberg Equilibrium test for SNPs on the chromosome X in 
female controls.
}
\examples{
 
}
\author{
Junfang Chen
}
