% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeQC.R
\name{removedSnpHWEauto}
\alias{removedSnpHWEauto}
\title{Hardy Weinberg Equilibrium test for autosomal SNPs}
\usage{
removedSnpHWEauto(groupLabel, plink, inputPrefix, pval, outputPvalFile,
  outputSNPfile, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{pval}{the p-value cutoff for controlling HWE test in either control or 
case subjects. Only autosomal SNPs are considered.}

\item{outputPvalFile}{the output pure text file that stores autosomal SNPs and 
their sorted HWE p-values.}

\item{outputSNPfile}{the output pure text file that stores the removed SNPs, 
one per line.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files.
}
\description{
Remove autosomal SNPs deviating from Hardy Weinberg Equilibrium (HWE).
}
\examples{
 
}
\author{
Junfang Chen
}
