% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasingImpute.R
\name{removedSnpMissPostImp}
\alias{removedSnpMissPostImp}
\title{Remove SNPs after post imputation}
\usage{
removedSnpMissPostImp(plink, inputPrefix, missCutoff, outputSNPfile,
  outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working 
directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{missCutoff}{the cutoff of the least number of instances for 
a SNP that is not missing.}

\item{outputSNPfile}{the output file of SNPs with pre-defined 
missing values.}

\item{outputPrefix}{the prefix of the PLINK binary files.}
}
\value{
The PLINK binary files after post imputation quality control 
and a pure text file contains SNPs with pre-defined missing values.
}
\description{
Remove SNPs which have a non missing value for less than a predefined 
number of instances.
}
\author{
Junfang Chen
}
