% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removedUnmapProbes.R
\name{removedUnmapProbes}
\alias{removedUnmapProbes}
\title{Remove unmapped probes or SNPs}
\usage{
removedUnmapProbes(plink, inputPrefix, chipAnnoFile, outputPrefix,
  outputSNPunmapFile)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{chipAnnoFile}{a pure text file that stores the chip annotation information (Affymetrix, Illumination, PsychChip and so on). 
This file can be found http://www.well.ox.ac.uk/~wrayner/strand/.}

\item{outputPrefix}{the prefix of the output PLINK format files.}

\item{outputSNPunmapFile}{a pure text file that stores the probe/SNP IDs, which are not mapped to the chip annotation file.}
}
\value{
The output PLINK format files after removing unmapped probe IDs.
}
\description{
Remove probes or SNPs that are not mapped to the chip annotation information. This chip annotation file is defined in the configuration folder.
}
\author{
Junfang Chen <junfang.chen3@gmail.com>
}
