% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{removeDupID}
\alias{removeDupID}
\title{Remove duplicated sample IDs}
\usage{
removeDupID(plink, dupSampleIDFile, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{dupSampleIDFile}{a pure text file that stores the duplicated sample IDs, 
each ID per line. If it is null, then duplicate the input PLINK files from 
the last step as the output files.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files after removing duplicated sample IDs.
}
\description{
Remove duplicated sample IDs from PLINK binary files. The duplicated IDs
are defined in a plain text file in advance.
}
\author{
Junfang Chen
}
