% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{removedYMtSnp}
\alias{removedYMtSnp}
\title{Remove SNPs on the chromosome Y and mitochondria}
\usage{
removedYMtSnp(plink, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{outputPrefix}{the prefix of the output PLINK format files.}
}
\value{
The output PLINK format files after splitting chromosome X into pseudoautosomal region and non-pseudoautosomal region.
}
\description{
Remove SNPs on the chromosome Y and mitochondria (in this case, chromosome==24 & chromosome==26 are removed). If no such SNPs, the output files remain the same as the input.
}
\author{
Junfang Chen <junfang.chen@zi-mannheim.de>
}
