% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{removeNoGroupId}
\alias{removeNoGroupId}
\title{Remove samples without group information}
\usage{
removeNoGroupId(plink, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}

\item{metaDataFile}{a pure text file that stores the meta information of 
the samples.}
}
\value{
The output PLINK binary files after removing samples without group 
information.
}
\description{
Remove samples without group/outcome/phenotype information, which is coded 
as -9 in the PLINK .FAM file.
}
\author{
Junfang Chen
}
