% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{removeSampID}
\alias{removeSampID}
\title{Remove samples in PLINK files}
\usage{
removeSampID(plink, removedSampIDFile, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{removedSampIDFile}{a pure text file that stores the useless 
sample IDs, each ID per line. If it is null, then duplicate the input 
PLINK files as the output files.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files after removing certain sample IDs.
}
\description{
Remove sample IDs that are useless such as duplicated or related IDs from 
PLINK binary files. These IDs are defined in a plain text file.
}
\examples{
## In the current working directory
bedFile <- system.file("extdata", "controlData.bed", package="Gimpute")
bimFile <- system.file("extdata", "controlData.bim", package="Gimpute") 
famFile <- system.file("extdata", "controlData.fam", package="Gimpute")
system(paste0("scp ", bedFile, bimFile, famFile, " ."))  
inputPrefix <- "controlData" ## Specify the input PLINK file prefix
removedSampIDFile <- system.file("extdata", "excludedSampIDs.txt", package="Gimpute")
outputPrefix <- "1_02_removedExclInst" 
## Not run: Requires an executable program PLINK, e.g.
## plink <- "/home/tools/plink"
## removeSampID(plink, removedSampIDFile, inputPrefix, outputPrefix)
}
\author{
Junfang Chen
}
