% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamedSnpChr.R
\name{renamedSnpChr}
\alias{renamedSnpChr}
\title{Remaming chromosomes}
\usage{
renamedSnpChr(plink, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{outputPrefix}{the prefix of the output PLINK format files.}
}
\value{
The output PLINK format files after renaming chromosomes.
}
\author{
Junfang Chen <junfang.chen3@gmail.com>
}
