% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managePlinkData.R
\name{renamePlinkBFile}
\alias{renamePlinkBFile}
\title{Rename PLINK binary files}
\usage{
renamePlinkBFile(inputPrefix, outputPrefix, action)
}
\arguments{
\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
Renamed PLINK binary files.
}
\description{
Rename a set of PLINK binary files (.BED, .BIM and .FAM).
}
\details{
The original input files can be retained using the action "copy"
or removed by using "move".
}
\author{
Junfang Chen
}
