% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{replaceGroupIdAndSex}
\alias{replaceGroupIdAndSex}
\title{Replace group and geneder presentation into proper PLINK format}
\usage{
replaceGroupIdAndSex(plink, inputPrefix, metaDataFile, outputPrefix)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{metaDataFile}{a pure text file that stores the meta information of the samples.}

\item{outputPrefix}{the prefix of the output PLINK format files.}
}
\value{
The output PLINK format files after presenting gender/group/outcome information into proper PLINK format.
}
\description{
Replace group and gender presentation into proper and consistent PLINK format.
}
\author{
Junfang Chen <junfang.chen@zi-mannheim.de>
}
