% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeQC.R
\name{setHeteroHaploMissing}
\alias{setHeteroHaploMissing}
\title{Set haploid heterozygous SNPs as missing}
\usage{
setHeteroHaploMissing(plink, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files after setting haploid heterozygous 
SNPs as missing.
}
\description{
Set all heterozygous alleles of chromosome X SNPs in male as missing.
}
\examples{
## In the current working directory
bedFile <- system.file("extdata", "genoUpdatedData.bed", package="Gimpute")
bimFile <- system.file("extdata", "genoUpdatedData.bim", package="Gimpute") 
famFile <- system.file("extdata", "genoUpdatedData.fam", package="Gimpute")
system(paste0("scp ", bedFile, bimFile, famFile, " ."))  
inputPrefix <- "genoUpdatedData" 
outputPrefix <- "2_03_setHeteroHaploMissing" 
## Not run: Requires an executable program PLINK, e.g.
## plink <- "/home/tools/plink"
## setHeteroHaploMissing(plink, inputPrefix, outputPrefix)
}
\author{
Junfang Chen
}
