% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{splitXchr}
\alias{splitXchr}
\title{Split chromosome X into pseudoautosomal region and non-pseudoautosomal region.}
\usage{
splitXchr(plink, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files after splitting chromosome X into 
pseudoautosomal region and non-pseudoautosomal region.
}
\description{
Split chromosome X into pseudoautosomal region and non-pseudoautosomal region, 
if chromosome X data is available.
}
\details{
Genomic coordinate system is on genome build hg19.
}
\author{
Junfang Chen
}
