% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{updatedSnpInfo}
\alias{updatedSnpInfo}
\title{Update the SNP information}
\usage{
updatedSnpInfo(plink, inputPrefix, chipAnnoFile, chipType, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{chipAnnoFile}{a pure text file that stores the chip annotation 
information. 
If the chip annotation file is not available for your study, it can be 
downloaded from http://www.well.ox.ac.uk/~wrayner/strand/.}

\item{chipType}{a string name defines the type of the chip annotation file: 
'illumina', 'affymetrix' or 'psychChip'.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files after updating SNP information.
}
\description{
Update SNP information including SNP name, base-pair position, 
chromosomal location and the strand information.
}
\details{
The SNP information in the chip annotation file is used as
the reference.
}
\author{
Junfang Chen
}
