% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrWiseSplit.R
\name{chrWiseSplit}
\alias{chrWiseSplit}
\title{Split genome-wide genotyping data into separate files by chromosome}
\usage{
chrWiseSplit(plink, inputPrefix, chrX_PAR1suffix, chrX_PAR2suffix)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input plink files before splitting.}

\item{chrX_PAR1suffix}{if chromosome 25 is available and with PAR1, then generate the suffix with X_PAR1 for chrX_PAR1}

\item{chrX_PAR2suffix}{if chromosome 25 is available and with PAR2, then generate the suffix with X_PAR2 for chrX_PAR2}

\item{outputPrefix}{the prefix of the output plink files after splitting and the chromosome number will be appended separately.}
}
\value{
The output plink files for each chromosome and possibly also the suffix of chrX_PAR.
}
\description{
Split the whole genome genotyping data chromosome-wise; allow parallel computating for all chromosomes.
if chromosome 25 is also available, further split chr25 (PAR or Chr_XY) into PAR1 and PAR2 according to the genomic coordination GRCh37 
from https://en.wikipedia.org/wiki/Pseudoautosomal_region.
The locations of the PARs within GRCh37 are:  
PAR1	X	60001	2699520 
PAR2	X	154931044	155260560
}
\author{
Junfang Chen <junfang.chen3@gmail.com>
}
