% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasingImpute.R
\name{chunk4eachChr}
\alias{chunk4eachChr}
\title{Chunk each chromosome into multiple segments}
\usage{
chunk4eachChr(inputPrefix, outputPrefix, chrs, windowSize = 3e+06)
}
\arguments{
\item{inputPrefix}{the prefix of the input PLINK binary files for 
each chromosome.}

\item{outputPrefix}{the prefix of the output pure text files that keep 
all the chunks for each chromosome separately.}

\item{chrs}{specifiy the chromosome codes for chunking.}

\item{windowSize}{the window size of each chunk. 
The default value is 3000000.}
}
\value{
The output pure text files include all the chunks 
for each chromosome separately.
}
\description{
Chunk each chromosome genotyping data into multiple segments 
by a predefined window size.
}
\examples{
 
## In the current working directory
bedFile <- system.file("extdata", "controlData.bed", package="Gimpute")
bimFile <- system.file("extdata", "controlData.bim", package="Gimpute") 
famFile <- system.file("extdata", "controlData.fam", package="Gimpute")
system(paste0("scp ", bedFile, bimFile, famFile, " ."))   
inputPrefix <- "controlData"
outputPrefix <- "chunks_chr"
bimCurrent <- read.table(file=famFile, stringsAsFactors=FALSE)  
chrs <- names(table(bimCurrent[,1]))
print(chrs)   
chunk4eachChr(inputPrefix, outputPrefix, chrs, windowSize)
}
\author{
Junfang Chen
}
