% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasingImpute.R
\name{filterImputeData}
\alias{filterImputeData}
\title{Filter genetic variants}
\usage{
filterImputeData(plink, suffix4impute2info, impute2infoFile, infoScore,
  badImputeSNPfile, prefix4imputedPlink, prefix4imputedFilterPlink)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{suffix4impute2info}{the suffix of IMPUTE2 generated files that store the imputation quality score 
for each variant, i.e. .impute2_info files.}

\item{impute2infoFile}{the output file of impute2 info scores which consists of two columns, namely all imputed SNPs and their info scores. 
Note that imputed SNPs with more than two alleles are dropped.}

\item{infoScore}{the cutoff of filtering imputation quality score for each variant.}

\item{badImputeSNPfile}{the output file of SNPs with bad info scores.}

\item{prefix4imputedPlink}{the prefix of the final imputed PLINK format files.}

\item{prefix4imputedFilterPlink}{the prefix of the final imputed and filtered PLINK format files.}
}
\value{
The filtered PLINK format imputed files, a pure text file with all removed SNPs (with bad infoScore) 
and a pure text file with the info scores of all imputed SNPs that has two columns: the SNPs and the corresponding info scores.
}
\description{
Filter genetic variants accoring to the imputation quality score with the help of .impute2_info files generated by IMPUTE2. 
In this case we keep variants with imputation info score of greater than 0.6.
}
\examples{
  
}
\author{
Junfang Chen <junfang.chen@zi-mannheim.de>
}
