% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeQC.R
\name{genoQC}
\alias{genoQC}
\title{Quality control for genotype data}
\usage{
genoQC(plink, inputPrefix, snpMissCutOffpre = 0.05, sampleMissCutOff = 0.02,
  Fhet = 0.2, snpMissCutOffpost = 0.02, snpMissDifCutOff = 0.02,
  femaleChrXmissCutoff = 0.05, pval4autoCtl = 1e-06,
  pval4femaleXctl = 1e-06, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{snpMissCutOffpre}{the cutoff of the missingness for removing SNPs 
before subject removal. The default is 0.05.}

\item{sampleMissCutOff}{the cutoff of the missingness for removing 
subjects/instances. The default is}

\item{Fhet}{the cutoff of the autosomal heterozygosity deviation. 
The default is 0.2.}

\item{snpMissCutOffpost}{the cutoff of the missingness for removing SNPs 
after subject removal. The default is 0.02.}

\item{snpMissDifCutOff}{the cutoff of the difference in missingness between 
cases and controls. The default is 0.02.}

\item{femaleChrXmissCutoff}{the cutoff of the missingness in female 
chromosome X SNPs. The default is 0.05.}

\item{pval4autoCtl}{the p-value cutoff for controlling HWE test in either 
control or case subjects. Only autosomal SNPs are considered. 
The default is 0.000001}

\item{pval4femaleXctl}{the p-value cutoff for controlling HWE test in 
female control subjects. Only chromosome X SNPs are considered. 
The default is 0.000001}

\item{outputPrefix}{the prefix of the output PLINK binary files after QC.}
}
\value{
The output PLINK binary files after QC.
}
\description{
Perform quality control on the genotype data.
}
\details{
The original PLINK files are implicitly processed by the following 
default steps: 
1.) Set all heterozygous alleles of SNPs on male chrX as missing;
2.) SNP missingness < 0.05 (before sample removal);
3.) Subject missingness < 0.02;  
4.) Remove subjects with |Fhet| >= 0.2;
5.) Reset paternal and maternal codes;
6.) SNP missingness < 0.02 (after sample removal);
7.) Remove SNPs with difference >= 0.02 of SNP missingness 
between cases and controls;
8.) Remove chrX SNPs with missingness >= 0.05 in females.
(Optional, if no chrX data);
9.) Remove autosomal SNPs with HWE p < 10-6 in controls;
10.) Remove chrX SNPs with HWE p < 10-6 in female controls. 
(Optional, if no chrX data).
}
\examples{
 
}
\author{
Junfang Chen
}
