% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeQC.R
\name{plotPCA4plink}
\alias{plotPCA4plink}
\title{Population outlier detection via PCA}
\usage{
plotPCA4plink(gcta, inputPrefix, outputPC4subjFile, outputPCplotFile)
}
\arguments{
\item{gcta}{an executable GCTA program in either the current working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{outputPC4subjFile}{the pure text file that stores all the subject IDs after QC and their corresponding eigenvalues of the first two principle components.}

\item{outputPCplotFile}{the plot file for visualizing the first two principle components of all subjects.}
}
\value{
The output pure text file and plot file for storing first two principle components of study subjects.
}
\description{
Perform principle component analysis (PCA) on quality controlled genotype data set in order to detect population outliers, 
and plot the PCs for visualization. Note that, only autosomal genotypes are used for computing principle components.
}
\author{
Junfang Chen <junfang.chen@zi-mannheim.de>
}
