% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{prepareChipAnnoFile4affymetrix}
\alias{prepareChipAnnoFile4affymetrix}
\title{prepare Affymetrix chip annotation file}
\usage{
prepareChipAnnoFile4affymetrix(inputFile, outputFile)
}
\arguments{
\item{inputFile}{an input pure text file that shows the chip annotation file which can be downloaded from http://www.well.ox.ac.uk/~wrayner/strand/.}

\item{outputFile}{an output pure text file that stores the chip annotation information in used-defined format (Affymetrix, Illumination, PsychChip and so on).}
}
\value{
a pure text file that stores the prepared chip annotation information in user-defined format.
}
\description{
Prepare Affymetrix chip annotation file into the format of interest.
}
\author{
Junfang Chen <junfang.chen@zi-mannheim.de>
}
