% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{removedDoubleProbes}
\alias{removedDoubleProbes}
\title{Remove duplicated SNPs}
\usage{
removedDoubleProbes(plink, inputPrefix, chipAnnoFile, chipType,
  outputSNPdupFile, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{chipAnnoFile}{an input chip annotation file. 
If the chip annotation file is not available for your study, it can be 
downloaded from http://www.well.ox.ac.uk/~wrayner/strand/.}

\item{chipType}{a string name that defines the type of the chip annotation 
file: 'illumina', 'affymetrix' or 'psychChip'.}

\item{outputSNPdupFile}{a pure text file that stores the duplicated SNP IDs, 
which are detected by the use of the chip annotation file.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files after removing duplicated SNP IDs.
}
\description{
Remove duplicated SNPs that have same rs-names or duplicated genomic position.
}
\details{
Duplicated SNPs have two levels of meaning: 1.) SNPs have same 
rs-names but different versions of SNP ID ound in chip annotation file. 
e.g. SNP-A IDs for Affymetrix chip. 2.) SNPs with duplicated genomic 
position: the combination of base pair position and chromosomal location.
}
\examples{
## In the current working directory
bedFile <- system.file("extdata", "controlData.bed", package="Gimpute")
bimFile <- system.file("extdata", "controlData.bim", package="Gimpute") 
famFile <- system.file("extdata", "controlData.fam", package="Gimpute")
chipAnnoFile <- system.file("extdata", "chipAnno.txt", package="Gimpute")
system(paste0("scp ", bedFile, bimFile, famFile, " ."))
inputPrefix <- "controlData"     
chipType <- "affymetrix"
outputSNPdupFile <- "snpDup.txt"
outputPrefix <- "removedDoubleProbes"
## Not run: Requires an executable program PLINK, e.g.
## plink <- "/home/tools/plink"
## removedDoubleProbes(plink, inputPrefix, chipAnnoFile,
##                     chipType, outputSNPdupFile, outputPrefix)
}
\author{
Junfang Chen
}
