% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removedExclProbe.R
\name{removedExclProbe}
\alias{removedExclProbe}
\title{Remove probes or SNPs}
\usage{
removedExclProbe(plink, inputPrefix, excludedProbeIdsFile, outputPrefix)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{excludedProbeIdsFile}{a pure text file that stores the probe IDs, which should be removed. 
If it is null, then copy and paste the input PLINK files from the last step.}

\item{outputPrefix}{the prefix of the output PLINK format files.}
}
\value{
The output PLINK format files after removing unwanted probe IDs.
}
\description{
Remove probes or SNPs that may be duplicated, or with unexpected probe names, which should be defined in advance. 
For instance, one can define these probes in the configuration folder.
}
\author{
Junfang Chen <junfang.chen3@gmail.com>
}
