% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{removedExclProbe}
\alias{removedExclProbe}
\title{Remove improper SNPs}
\usage{
removedExclProbe(plink, inputPrefix, excludedProbeIdsFile, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{excludedProbeIdsFile}{a pure text file that stores the SNP IDs, 
one per line, which need to be removed. If it is null, then no SNPs are removed.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files after removing unwanted SNP IDs.
}
\description{
Remove SNPs that may be duplicated, or with unexpected SNP names.
}
\details{
excludedProbeIdsFile should be defined in a plain text file 
in advance. Improper SNPs such as AFFX, cnvi etc.. with unexpect format
must be excluded.
}
\examples{
## In the current working directory
bedFile <- system.file("extdata", "controlData.bed", package="Gimpute")
bimFile <- system.file("extdata", "controlData.bim", package="Gimpute") 
famFile <- system.file("extdata", "controlData.fam", package="Gimpute")
excludedProbeIdsFile <- system.file("extdata", "excludedProbeIDs.txt", package="Gimpute")
system(paste0("scp ", bedFile, bimFile, famFile, " ."))
inputPrefix <- "controlData" ## Specify the input PLINK file prefix
outputPrefix <- "1_06_removedExclProbe" 
## Not run: Requires an executable program PLINK, e.g.
## plink <- "/home/tools/plink"
## removedExclProbe(plink, inputPrefix, excludedProbeIdsFile, outputPrefix)
}
\author{
Junfang Chen
}
