% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removedInstFhet.R
\name{removedInstFhet}
\alias{removedInstFhet}
\title{Remove subjects with great autosomal heterozygosity deviation}
\usage{
removedInstFhet(plink, Fhet, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{Fhet}{the cutoff of the autosomal heterozygosity deviation.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{outputPrefix}{the prefix of the output PLINK format files.}
}
\value{
The output PLINK format files after removing subjects with great autosomal heterozygosity deviation.
}
\description{
Remove instances with great autosomal heterozygosity deviation. i.e. with |Fhet| >= 0.2.
This analysis will automatically skip haploid markers (male X and Y chromosome markers).
}
\author{
Junfang Chen <junfang.chen3@gmail.com>
}
