% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeQC.R
\name{removedInstMiss}
\alias{removedInstMiss}
\title{Remove subjects with missing values}
\usage{
removedInstMiss(plink, sampleMissCutOff, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{sampleMissCutOff}{the cutoff of the missingness for removing 
subjects/instances.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files after removing subjects with  
a pre-defined removal cutoff.
}
\description{
Remove Subjects or instances with missingness of greater than a certain 
threshold.
}
\examples{
 
}
\author{
Junfang Chen
}
