% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removedMonoSnp.R
\name{removedMonoSnp}
\alias{removedMonoSnp}
\title{Remove monomorphic SNPs}
\usage{
removedMonoSnp(plink, inputPrefix, outputPrefix, outputSNPs)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input plink files.}

\item{outputPrefix}{the prefix of the output plink files (after removing monomorphic SNPs).}

\item{outputSNPs}{the output pure text file that stores the removed monomorphic SNPs.}
}
\value{
The output plink files (after removing monomorphic SNPs) and a pure text file with removed monomorphic SNPs.
}
\description{
Remove monomorphic SNPs after QC and alignment to the imputation reference.
}
\author{
Junfang Chen <junfang.chen3@gmail.com>
}
