% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasingImpute.R
\name{removedMonoSnp}
\alias{removedMonoSnp}
\title{Exclude monomorphic SNPs}
\usage{
removedMonoSnp(plink, inputPrefix, outputPrefix, outputSNPfile)
}
\arguments{
\item{plink}{an executable program in either the current 
working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{outputPrefix}{the prefix of the output PLINK binary files 
after removing monomorphic SNPs.}

\item{outputSNPfile}{the output pure text file that stores 
the removed monomorphic SNPs, one per line, if any.}
}
\value{
The output PLINK binary files after removing monomorphic SNPs 
and a pure text file with removed monomorphic SNPs.
}
\description{
Detect monomorphic SNPs from PLINK BIM file and exclude them if any.
}
\examples{
 
## In the current working directory
bedFile <- system.file("extdata", "alignedData.bed", package="Gimpute")
bimFile <- system.file("extdata", "alignedData.bim", package="Gimpute") 
famFile <- system.file("extdata", "alignedData.fam", package="Gimpute")
system(paste0("scp ", bedFile, " ."))   
system(paste0("scp ", bimFile, " ."))   
system(paste0("scp ", famFile, " ."))   
inputPrefix <- "alignedData"
outputPrefix <- "removedMonoSnp" 
outputSNPfile <- "monoSNP.txt"  
## Not run: Requires an executable program PLINK, e.g.
## plink <- "/home/tools/plink"
## removedMonoSnp(plink, inputPrefix, outputPrefix, outputSNPfile)
}
\author{
Junfang Chen
}
