% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removedParentIdsMiss.R
\name{removedParentIdsMiss}
\alias{removedParentIdsMiss}
\title{Reset paternal and maternal codes}
\usage{
removedParentIdsMiss(plink, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{outputPrefix}{the prefix of the output PLINK format files.}
}
\value{
The output PLINK format files.
}
\description{
Reset paternal and maternal codes of non-founders if parents not present. 
Replace the paternal ID and maternal ID of instances (childs) by the
value zero if the paternal ID and the maternal ID do not belong to any
instance (parent) with the same family ID as the child.
}
\author{
Junfang Chen <junfang.chen3@gmail.com>
}
