% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeQC.R
\name{removedSnpFemaleChrXmiss}
\alias{removedSnpFemaleChrXmiss}
\title{remove chromosome X SNPs in females}
\usage{
removedSnpFemaleChrXmiss(plink, femaleChrXmissCutoff, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{femaleChrXmissCutoff}{the cutoff of the missingness 
in female chromosome X SNPs.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files.
}
\description{
Remove SNPs on the chromosome X with a pre-defined cutoff for 
missingness in females.
}
\examples{
 
}
\author{
Junfang Chen
}
