% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeQC.R
\name{removedSnpHetX}
\alias{removedSnpHetX}
\title{Remove heterozygous SNPs in haploid male chromosome X only if chromosome X data exists.}
\usage{
removedSnpHetX(plink, inputPrefix, hhCutOff, outputPrefix,
  outputFile_SNPhhFreqAll, outputFile_SNPhhFreqRetained)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{hhCutOff}{the cutoff for removing male haploid heterozygous SNPs on the chromosome X.}

\item{outputPrefix}{the prefix of the output PLINK format files.}

\item{outputFile_SNPhhFreqAll}{the output pure text file that stores all heterozygous SNPs with their frequency (the number of males for this SNP), if any.}

\item{outputFile_SNPhhFreqRetained}{the output pure text file that stores retained heterozygous SNPs with their frequency (the number of males for this SNP), if any.}
}
\value{
The output PLINK format files and two pure text files (if any) with heterozygous SNPs and their respective frequency.
}
\description{
In principle, one has to remove all heterozygous SNPs of chromosome X in males. However, too many SNPs might be removed in some data sets. 
We want to accept a small percentage of such SNPs in the data set so that we do not lose too much SNPs.
}
\author{
Junfang Chen <junfang.chen@zi-mannheim.de>
}
