% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removedSnpHWEautoControl.R
\name{removedSnpHWEautoControl}
\alias{removedSnpHWEautoControl}
\title{Hardy weinberg equilibrium test for autosomal SNPs in controls.}
\usage{
removedSnpHWEautoControl(plink, inputPrefix, pval, outputFile_pVal,
  outputSNPfile, outputPrefix)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{pval}{the p-value cutoff for controlling HWE test in control subjects. Only autosomal SNPs are tested here.}

\item{outputFile_pVal}{the output pure text file that stores autosomal SNPs and their sorted HWE p-values.}

\item{outputSNPfile}{the output pure text file that stores the removed SNPs.}

\item{outputPrefix}{the prefix of the output PLINK format files.}
}
\value{
The output PLINK format files.
}
\description{
Remove autosomal SNPs in controls.
}
\author{
Junfang Chen <junfang.chen3@gmail.com>
}
