% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeQC.R
\name{removedSnpMiss}
\alias{removedSnpMiss}
\title{Remove SNPs with missing values}
\usage{
removedSnpMiss(plink, snpMissCutOff, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{snpMissCutOff}{the cutoff of the missingness for removing SNPs.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files after removing SNPs with pre-defined 
missing values.
}
\description{
Remove SNPs with missingness of greater than a certain threshold.
}
\examples{
 
## In the current working directory
bedFile <- system.file("extdata", "genoUpdatedData.bed", package="Gimpute")
bimFile <- system.file("extdata", "genoUpdatedData.bim", package="Gimpute") 
famFile <- system.file("extdata", "genoUpdatedData.fam", package="Gimpute")
system(paste0("scp ", bedFile, bimFile, famFile, " ."))  
snpMissCutOff <- 0.05
inputPrefix <- "genoUpdatedData" 
outputPrefix <- "2_04_removedSnpMissPre" 
## Not run: Requires an executable program PLINK, e.g.
## plink <- "/home/tools/plink"
## removedSnpMiss(plink, snpMissCutOff, inputPrefix, outputPrefix)
}
\author{
Junfang Chen
}
