% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removedSnpMiss.R
\name{removedSnpMiss}
\alias{removedSnpMiss}
\title{Remove SNPs with missing values}
\usage{
removedSnpMiss(plink, snpMissCutOff, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{snpMissCutOff}{the cutoff of the missingness for removing SNPs.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{outputPrefix}{the prefix of the output PLINK format files.}
}
\value{
The output PLINK format files after removing SNPs with pre-defined missing values.
}
\description{
Remove SNPs with missingness of greater than a certain threshold before/after removing subjects.
}
\author{
Junfang Chen <junfang.chen3@gmail.com>
}
