% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removedSnpMissDiff.R
\name{removedSnpMissDiff}
\alias{removedSnpMissDiff}
\title{Remove SNPs with difference in SNP missingness between cases and controls.}
\usage{
removedSnpMissDiff(plink, inputPrefix, snpMissDifCutOff, outputPrefix,
  caseControl)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{snpMissDifCutOff}{the cutoff of the difference in missingness between cases and controls.}

\item{outputPrefix}{the prefix of the output PLINK format files.}

\item{caseControl}{a logical value indicating whether the data set contains both case-control subjects.}
}
\value{
The output PLINK format files.
}
\description{
Remove SNPs with difference in SNP missingness between cases and controls. 
To test for differential call rates between cases and controls for each SNP
This only works for the genotype data set with cases and controls.
}
\author{
Junfang Chen <junfang.chen3@gmail.com>
}
