% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasingImpute.R
\name{removedSnpMissPostImp}
\alias{removedSnpMissPostImp}
\title{Remove SNPs after post imputation}
\usage{
removedSnpMissPostImp(plink, inputPrefix, missCutoff, outRemovedSNPfile,
  outRetainSNPfile, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working 
directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{missCutoff}{the cutoff of the least number of instances for 
a SNP that is not missing. The default is 20.}

\item{outRemovedSNPfile}{the output file of SNPs with pre-defined 
missing values that are removed.}

\item{outRetainSNPfile}{the output file of SNPs that are retained.}

\item{outputPrefix}{the prefix of the PLINK binary files.}
}
\value{
The PLINK binary files after post imputation quality control 
and a pure text file contains SNPs with pre-defined missing values.
}
\description{
Remove SNPs which have a non missing value for less than a predefined 
number of instances.
}
\examples{
## In the current working directory
bedFile <- system.file("extdata", "alignedData.bed", package="Gimpute")
bimFile <- system.file("extdata", "alignedData.bim", package="Gimpute") 
famFile <- system.file("extdata", "alignedData.fam", package="Gimpute")
system(paste0("scp ", bedFile, " ."))   
system(paste0("scp ", bimFile, " ."))   
system(paste0("scp ", famFile, " ."))    
inputPrefix <- "alignedData" 
outputPrefix <- "removedSnpMissPostImp" 
## Not run: Requires an executable program PLINK, e.g.
## plink <- "/home/tools/plink"
## removedSnpMissPostImp(plink, inputPrefix, missCutoff=20, 
##                       outRemovedSNPfile, outRetainSNPfile, outputPrefix)
## 
}
\author{
Junfang Chen
}
