% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{removedUnmapProbes}
\alias{removedUnmapProbes}
\title{Remove SNPs not in the chip annotation file}
\usage{
removedUnmapProbes(plink, inputPrefix, chipAnnoFile, outputPrefix,
  outputSNPfile)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{chipAnnoFile}{a pure text file that stores the chip annotation
information.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}

\item{outputSNPfile}{a pure text file that stores the SNP IDs, 
one per line, which are not mapped to the chip annotation file.}
}
\value{
The output text file contains the removed SNP IDs, one per line.
The PLINK binary files after removing unmapped SNP IDs.
}
\description{
Check if all SNPs are included in the chip annotation file. 
If some of the input SNPs are not included, then remove them.
}
\details{
If the chip annotation file is not available for your study, 
you can download it from http://www.well.ox.ac.uk/~wrayner/strand/.
}
\author{
Junfang Chen
}
