% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{removedUnmapProbes}
\alias{removedUnmapProbes}
\title{Remove SNPs not in the chip annotation file}
\usage{
removedUnmapProbes(
  plink,
  inputPrefix,
  chipAnnoFile,
  chipType,
  outputPrefix,
  outputSNPfile
)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{chipAnnoFile}{a pure text file that stores the chip annotation
information.}

\item{chipType}{a string name defines the type of the chip annotation file: 
'SNPIDstudy', and 'rsIDstudy'. The detail is described in 
\code{\link{prepareAnnoFile4affy}}.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}

\item{outputSNPfile}{a pure text file that stores the SNP IDs, 
one per line, which are not mapped to the chip annotation file.}
}
\value{
The output text file contains the removed SNP IDs, one per line.
The PLINK binary files after removing unmapped SNP IDs.
}
\description{
Check if all SNPs are included in the chip annotation file. 
If some of the input SNPs are not included, then remove them.
}
\details{
If the chip annotation file is not available for your study, 
you can download it from http://www.well.ox.ac.uk/~wrayner/strand/.
}
\examples{
## In the current working directory
bedFile <- system.file("extdata", "controlData.bed", package="Gimpute")
bimFile <- system.file("extdata", "controlData.bim", package="Gimpute") 
famFile <- system.file("extdata", "controlData.fam", package="Gimpute")
chipAnnoFile <- system.file("extdata", "chipAnno.txt", package="Gimpute")
system(paste0("scp ", bedFile, bimFile, famFile, " ."))
inputPrefix <- "controlData" ## Specify the input PLINK file prefix
outputSNPfile <- "1_07_removedUnmapProbes"   
## Not run: Requires an executable program PLINK, e.g.
## plink <- "/home/tools/plink"
## removedUnmapProbes(plink, inputPrefix, chipAnnoFile, chipType, 
##                    outputPrefix, outputSNPfile)
}
\seealso{
\code{\link{prepareAnnoFile4affy}}
}
\author{
Junfang Chen
}
