% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{removeDupID}
\alias{removeDupID}
\title{Remove duplicated sample IDs}
\usage{
removeDupID(plink, dupSampleIDFile, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{dupSampleIDFile}{a pure text file that stores the duplicated sample IDs, each ID per line. 
If it is null, then copy and paste the input PLINK files from the last step as the output files.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{outputPrefix}{the prefix of the output PLINK format files.}
}
\value{
The output PLINK format files after removing duplicated sample IDs.
}
\description{
Remove duplicated sample IDs, which should be defined in the configuration folder in advance.
}
\author{
Junfang Chen <junfang.chen@zi-mannheim.de>
}
