% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{removedWrongAnceInst}
\alias{removedWrongAnceInst}
\title{Remove samples with improper ancestry}
\usage{
removedWrongAnceInst(plink, inputPrefix, metaDataFile, ancestrySymbol,
  outputPrefix)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{metaDataFile}{a pure text file that stores the meta information of the samples.}

\item{ancestrySymbol}{an indicator that shows the symbol of targeted ancestry. Such as 'EA' stands for the European, 'AA' for African American. 
If it is null, then all samples are selected.}

\item{outputPrefix}{the prefix of the output PLINK format files.}
}
\value{
The output PLINK format files after ancestry check.
}
\description{
Remove samples with improper ancestry or keep samples with your own purpose.
}
\author{
Junfang Chen <junfang.chen@zi-mannheim.de>
}
