% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{removedYMtSnp}
\alias{removedYMtSnp}
\title{Remove SNPs on the chromosome Y and mitochondrial DNA}
\usage{
removedYMtSnp(plink, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files after removing SNPs 
on the chromosome Y and mitochondrial DNA.
}
\description{
Remove SNPs on the chromosome Y and mitochondrial DNA.
}
\details{
Note that if chromosome Y and mitochondrial DNA are available,
they must be coded as 24 and 26, respectively.
}
\author{
Junfang Chen
}
