% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{removedYMtSnp}
\alias{removedYMtSnp}
\title{Remove SNPs on the chromosome Y and mitochondrial DNA}
\usage{
removedYMtSnp(plink, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files after removing SNPs 
on the chromosome Y and mitochondrial DNA.
}
\description{
Remove SNPs on the chromosome Y and mitochondrial DNA.
}
\details{
Note that if chromosome Y and mitochondrial DNA are available,
they must be coded as 24 and 26, respectively.
}
\author{
Junfang Chen 
## In the current working directory
bedFile <- system.file("extdata", "controlData.bed", package="Gimpute")
bimFile <- system.file("extdata", "controlData.bim", package="Gimpute") 
famFile <- system.file("extdata", "controlData.fam", package="Gimpute")
system(paste0("scp ", bedFile, bimFile, famFile, " ."))  
inputPrefix <- "controlData" ## Specify the input PLINK file prefix 
outputPrefix <- "1_11_removedYMtSnp" 
## Not run: Requires an executable program PLINK, e.g.
## plink <- "/home/tools/plink"
## removedYMtSnp(plink, inputPrefix, outputPrefix)
}
