% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{removeNoGroupId}
\alias{removeNoGroupId}
\title{Remove samples without group information}
\usage{
removeNoGroupId(plink, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{outputPrefix}{the prefix of the output PLINK format files.}

\item{metaDataFile}{a pure text file that stores the meta information of the samples.}
}
\value{
The output PLINK format files after removing samples without group information/IDs.
}
\description{
Remove samples without group/outcome/phenotype information, which is coded as -9 in PLINK .fam file.
}
\author{
Junfang Chen <junfang.chen@zi-mannheim.de>
}
