% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeQC.R
\name{removeOutlierByPCs}
\alias{removeOutlierByPCs}
\title{Remove population outliers}
\usage{
removeOutlierByPCs(plink, gcta, inputPrefix, nThread, cutoff, cutoffSign,
  inputPC4subjFile, outputPC4outlierFile, outputPCplotFile, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{gcta}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{nThread}{the number of threads used for computation.}

\item{cutoff}{the cutoff that distinguishes the eigenvalues of the outliers
from ordinary population. If it is null, then there are no outliers or 
outliers are not required to be removed.}

\item{cutoffSign}{the cutoff sign: 'greater' or 'smaller' that determines 
if the outliers should be greater or smaller than the cutoff value.}

\item{inputPC4subjFile}{the pure text file that stores all the subject IDs 
and their corresponding eigenvalues of the first two principle components.}

\item{outputPC4outlierFile}{the pure text file that stores the outlier IDs 
and their corresponding eigenvalues of the first two principle components.}

\item{outputPCplotFile}{the plot file for visualizing the first two 
principle components of all subjects without population outliers.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
1.) The output PLINK binary files after outlier removal. 
2.) The output pure text file (if any) for storing removed outlier IDs 
and their corresponding PCs. 3.) The plot file (if any) for visualizing 
the first two principle components after outlier removal.
}
\description{
Remove population outliers by using principle component analysis.
}
\details{
This function is used for removing population outliers. 
If the outliers are necessary to be removed, then one uses the eigenvalues 
from the first principle component as a criterion to find out the outliers 
by assigning an appropriate cutoff.
}
\author{
Junfang Chen
}
