% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managePlinkData.R
\name{renamePlinkBFile}
\alias{renamePlinkBFile}
\title{Rename PLINK binary files}
\usage{
renamePlinkBFile(inputPrefix, outputPrefix, action)
}
\arguments{
\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}

\item{action}{a string indicating if the action is "copy" or "move".}
}
\value{
Renamed PLINK binary files.
}
\description{
Rename a set of PLINK binary files (.BED, .BIM and .FAM).
}
\details{
The original input files can be retained using the action "copy"
or removed by using "move".
}
\examples{
## In the current working directory
bedFile <- system.file("extdata", "controlData.bed", package="Gimpute")
bimFile <- system.file("extdata", "controlData.bim", package="Gimpute") 
famFile <- system.file("extdata", "controlData.fam", package="Gimpute")
system(paste0("scp ", bedFile, bimFile, famFile, " ."))  
inputPrefix <- "controlData" 
outputPrefix <- "dataCtl" 
outputPrefix <- "1_02_removedExclInst"  
## renamePlinkBFile(inputPrefix, outputPrefix, action="move")
}
\author{
Junfang Chen
}
