% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageData.R
\name{renamePlinkBFile}
\alias{renamePlinkBFile}
\title{Rename PLINK binary files}
\usage{
renamePlinkBFile(inputPrefix, outputPrefix, action)
}
\arguments{
\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{outputPrefix}{the prefix of the output PLINK format files.}
}
\value{
Renamed PLINK binary files.
}
\description{
Rename a set of PLINK format files. Namely, .BED, .BIM and .FAM
}
\details{
The original input files can be retained using the action "copy"
or removed by using "move".
}
\author{
Junfang Chen <junfang.chen@zi-mannheim.de>
}
