% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeQC.R
\name{setHeteroHaploMissing}
\alias{setHeteroHaploMissing}
\title{Set haploid heterozygous SNPs as missing}
\usage{
setHeteroHaploMissing(plink, inputPrefix, outputPrefix)
}
\arguments{
\item{plink}{an executable PLINK program in either the current working directory or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK format files.}

\item{outputPrefix}{the prefix of the output PLINK format files.}
}
\value{
The output PLINK format files.
}
\description{
Set all heterozygous alleles of chromosome X SNPs in male as missing.
}
\author{
Junfang Chen <junfang.chen@zi-mannheim.de>
}
