% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{updateGenoInfo}
\alias{updateGenoInfo}
\title{Update genotype information}
\usage{
updateGenoInfo(plink, inputPrefix, metaDataFile, removedSampIDFile,
  ancestrySymbol, excludedProbeIdsFile, chipAnnoFile, chipType, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{removedSampIDFile}{a pure text file that stores the duplicated sample IDs, 
each ID per line. If it is null, then duplicate the input PLINK files from 
the last step as the output files.}

\item{ancestrySymbol}{an indicator that shows the symbol of genetic ancestry. 
If it is null, then all samples are selected.}

\item{excludedProbeIdsFile}{a pure text file that stores the SNP IDs, 
one per line, which need to be removed. If it is null, no SNPs are removed.}

\item{chipAnnoFile}{a pure text file that stores the chip annotation
information.}

\item{chipType}{a string name that defines the type of the chip annotation 
file: 'illumina', 'affymetrix' or 'psychChip'.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files after genotype information remapping.
}
\description{
Update genotype data information of the original PLINK binary files involving  
subject metadata information remapping and SNP information rearrangement and 
conversion according to the annotation file.
}
\details{
The original PLINK files are implicitly processed by the following steps: 
1.) remove duplicated subjects;
2.) update group ID and sex information;
3.) remove not labelled subjects;
4.) remove subjects with wrong ancestry;
5.) remove incorrectly annotated SNPs;
6.) remove SNPs that are not in the annotation file;
7.) remove duplicated SNPs;
8.) update SNP genomic position and strand information;
9.) split chromosome X into pseudoautosomal region (PAR) and non-PAR;
10.) remove SNPs on the chromosome Y and mitochondrial DNA.
The metadata information file and the chip annotation file are used as the 
reference for the update.
If the chip annotation file is not available for your study, it can be 
downloaded from http://www.well.ox.ac.uk/~wrayner/strand/.
}
\author{
Junfang Chen
}
