% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeInfoUpdate.R
\name{updateGroupIdAndSex}
\alias{updateGroupIdAndSex}
\title{Update group and geneder information}
\usage{
updateGroupIdAndSex(plink, inputPrefix, metaDataFile, outputPrefix)
}
\arguments{
\item{plink}{an executable program in either the current working directory 
or somewhere in the command path.}

\item{inputPrefix}{the prefix of the input PLINK binary files.}

\item{metaDataFile}{a pure text file that stores the meta information of 
the samples.}

\item{outputPrefix}{the prefix of the output PLINK binary files.}
}
\value{
The output PLINK binary files after updating the gender and grouping 
information.
}
\description{
Replace group and gender information in the PLINK binary files by using 
the information from the metadata file.
}
\details{
Find the shared sample IDs between PLINK input files and metadata file. 
Use the information from the metadata file as the reference and update the group 
information/the outcome in the PLINK file. Group label should be 1 and 2. 
(1=unaff, 2=aff, 0=miss); missing phenotype will be indicated as -9.
}
\author{
Junfang Chen
}
